Worker: reset internal state if we fail contacting the WebUI
When we have Websocket disconnections the $host gets removed, resulting
in impossibility to update the status of the job to the WebUI.

Previously we used just to let the worker loop die in this case, this lead
the worker to send messages of the job status, even if worker is not
working on it. In that case the WebUI will reap the job after the timeout expired.
Now we need to be sure from the worker that the status sent to the
server is consistent, and not leak previous jobs information: with this change we
 reset the state of worker instead of leaving the state unpredictable.

See related progress issue: https://progress.opensuse.org/issues/25124